+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [1.2.0] - Add TransformControllerProp
+## [1.1.0] - Add helper methods to various props (animation.value, tab.index, etc)
+## [1.0.2] - Rename ValueChanged
 ## [1.0.1] - Update FocusNodeProp
 ## [1.0.0] - Simplify implementation
 ## [0.1.5+1] - Fixed missing RestorationIds, add support for debugFillProperties, add NotifiersBuilder